Do local write combining to avoid async http explosions
Pack muliple event messages into a single influxdb write, and only
do a single async http call at once.  This also adds a stats page
to see what devices send huge number of events, and maybe help find
out why influxdb is a perceived source of many problems.

Issues:

1. runIn() will reschedule all subsequent calls, so this requires a 1s pause
between requests to actually complete a call, but for a 1st pass I'm ok with
this.

2. All messages are time stamped with local time before sending to influxdb
since calls could be delayed.  This simple and easy, so who not.  This should
probably use the event time as returned by evt.getUnixTime(), but this isn't
documented as to what it is, but this should be dug into since it surely
would be more accurate?

3. Influxdb line protocol (/write) is generated progressively as the line is
formed, and this is messy and could be centralized.  For example
postToInfluxDB() appends the timestamp to a completed line.  Also
postToInfluxDBNextFromQueue() specifies the precision which is implicit in the
earlier new Date().getTime() call.