Update influxdb-logger.groovy
add an option to control whether attribute events where the value is not changed are filtered or not.  Default = not filtered (possibly a breaking change, but I think it's the right choice in general)

Necessary to not filter events with no value change for devices where the last value represents the average of the time period since the previous value.  This is necessary for correct "step-back interpolation", e.g. flow sensors.